Use dbus-monitor --pcap to capture traffic
Modern versions of D-Bus have a BecomeMonitor method which allows a
sufficiently-privileged user (ie root) to capture all messages on the
system bus. Modern versions of dbus-monitor use this method, and support
pcap output natively.

This allows us to use pkexec to escalate just the dbus-monitor process
when monitoring the system bus. (Currently the UI does not allow you to
monitor the system bus.) We will also be able to use Flatpak's
HostCommand method to escape the sandbox, rather than requiring full
session bus access for Bustle itself.

The downside is that we lose the initial state dump of all names on the
bus. This seems like an acceptable price to pay, though I think we can
improve the inference of name owners to make this less costly anyway.